fix: moved color extensions to a specific file and improved SipCredentialHeader to match Figma design.
