+++ TelnyxWebRTCDemo/Views/SipCredentialHeader.swift
@@ -21,52 +21,29 @@ struct SipCredentialHeader: View {
                         .foregroundColor(Color(hex: "#1D1D1D"))
                 }
             }
-            
-            Button(action: onAddProfile) {
-                Text("+ Add new profile")
-                    .font(.system(size: 14))
-                    .foregroundColor(Color(hex: "#1D1D1D"))
-                    .frame(maxWidth: .infinity)
-                    .padding(.vertical, 8)
-                    .background(Color(hex: "#F5F3E4"))
-                    .cornerRadius(16)
+            HStack {
+                Button(action: onAddProfile) {
+                    Text("+ Add new profile")
+                        .font(.system(size: 14))
+                        .foregroundColor(Color(hex: "#1D1D1D"))
+                        .frame(width: 150)
+                        .padding(.vertical, 8)
+                        .background(Color(hex: "#F5F3E4"))
+                        .cornerRadius(16)
+                }
+                Spacer()
             }
-            
-            Text(subtitle)
-                .font(.system(size: 12))
-                .foregroundColor(.gray)
+            HStack {
+                Text(subtitle)
+                    .font(.system(size: 12))
+                    .foregroundColor(.gray)
+                Spacer()
+            }
         }
-        .padding(.horizontal, 15)
+        .padding(.horizontal, 24)
         .padding(.vertical, 8)
         .background(Color(uiColor: .systemBackground))
     }
-}
-
-extension Color {
-    init(hex: String) {
-        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int: UInt64 = 0
-        Scanner(string: hex).scanHexInt64(&int)
-        let a, r, g, b: UInt64
-        switch hex.count {
-        case 3:
-            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6:
-            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8:
-            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            (a, r, g, b) = (255, 0, 0, 0)
-        }
-        self.init(
-            .sRGB,
-            red: Double(r) / 255,
-            green: Double(g) / 255,
-            blue:  Double(b) / 255,
-            opacity: Double(a) / 255
-        )
-    }
-}
 }
 
 #Preview {
